% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstat.R
\name{freqdist}
\alias{freqdist}
\title{Freeqdist Function fit gor a given data}
\usage{
freqdist(x, ci)
}
\arguments{
\item{x}{vector whose frequency disribution is required}

\item{ci}{number of class intervals}
}
\value{
none
}
\description{
Freeqdist Function fit gor a given data
}
\examples{
freqdist()
}
\keyword{freqdist}

