% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstat.R
\name{meansd}
\alias{meansd}
\title{This function calculates both mean and standard deviation}
\usage{
meansd(x)
}
\arguments{
\item{x}{vector whose mean and standard deviations are required}
}
\value{
none
}
\description{
This function calculates both mean and standard deviation
}
\examples{
meansd()
}
\keyword{meansd}

