% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstat.R
\name{resana}
\alias{resana}
\title{A Regression Residual Analysis}
\usage{
resana(obj)
}
\arguments{
\item{obj}{object obtained from the lm function}
}
\value{
none
}
\description{
A Regression Residual Analysis
}
\examples{
resana()
}
\keyword{resana}

