% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstats.R
\name{stat}
\alias{stat}
\title{Create histogram, boxplot, and Numerical summary}
\usage{
stat(x)
}
\arguments{
\item{x}{Numerical variable}
}
\description{
Basic statistics descriptions
}
\author{
Coffi Rodolphe Segbedji
}
