% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bstat.R
\name{trntest}
\alias{trntest}
\title{Extract Train and Test datasets for any data frame}
\usage{
trntest(df, n)
}
\arguments{
\item{df}{dataframe for which train and test data sets are to be extrcted}

\item{n}{number of observations}
}
\value{
list containing two data frames train and test
}
\description{
Extract Train and Test datasets for any data frame
}
\examples{
trntest()
}
\keyword{trntest}

